requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
distro==1.9.0
Flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
MarkupSafe==3.0.2
marshmallow>=3.0.0,<4.0.0
openai==1.68.2
pydantic==2.10.6
pydantic_core==2.27.2
pytest>=7.0.0,<8.0.0
python-dotenv==1.1.0
requests>=2.25.0,<3.0.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.13.0
Werkzeug==3.1.3